# ada-server/web/swagger/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Keepa API",
    "description" : "The Keepa API offers numerous endpoints.  Every request requires your API access key as a parameter. You can find and change your key in the keepa portal. All requests must be issued as a HTTPS GET and accept gzip encoding. If possible, use a Keep-Alive connection.  Multiple requests can be made in parallel to increase throughput.",
    "contact" : {
      "name" : "Keepa",
      "url" : "https://keepa.com/#!api",
      "email" : "info@keepa.com"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "https://virtserver.swaggerhub.com/magicCashew/keepa/1.0.0",
    "description" : "SwaggerHub API Auto Mocking"
  }, {
    "url" : "https://api.keepa.com/",
    "description" : "The Production API which is accessible only by https."
  } ],
  "tags" : [ {
    "name" : "product",
    "description" : "Retrieve the product for the specified ASIN and domain."
  }, {
    "name" : "category",
    "description" : "Retrieve categories using the node id and (optional) parent tree."
  } ],
  "paths" : {
    "/product" : {
      "get" : {
        "tags" : [ "product" ],
        "summary" : "Retrieve the product for the specified ASIN and domain.",
        "description" : "Retrieves the product object for the specified ASIN and domain. If our last update is older than one hour it will be automatically refreshed before delivered to you to ensure you get near to real-time pricing data.  You can request products via either their ASIN (preferred) or via UPC and EAN codes. You can not use both parameters, asin and code, in the same request. Keepa can not track Amazon Fresh and eBooks.",
        "operationId" : "product",
        "parameters" : [ {
          "name" : "key",
          "in" : "query",
          "description" : "Your private API key.",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "domain",
          "in" : "query",
          "description" : "Integer value for the Amazon locale you want to access. (1-com, 2-co.uk, 3-de, 4-fr, 5-co.jp, 6-ca, 7-cn, 8-it, 9-es, 10-in, 11-com.mx, 12-com.br, 13-com.au)",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "enum" : [ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13 ]
          }
        }, {
          "name" : "asin",
          "in" : "query",
          "description" : "The ASIN of the product you want to request. For batch requests a comma separated list of ASINs (up to one hundred).",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "code",
          "in" : "query",
          "description" : "The product code of the product you want to request. We currently allow UPC, EAN and ISBN-13 codes. For batch requests a comma separated list of codes (up to one hundred). Multiple ASINs can have the same product code, so requesting a product code can return multiple products.",
          "required" : false,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/category"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      }
    },
    "/category" : {
      "get" : {
        "tags" : [ "category" ],
        "summary" : "Returns Amazon category information from Keepa API.",
        "description" : "Retrieve category objects using their node ids and (optional) their parent tree.",
        "operationId" : "category",
        "parameters" : [ {
          "name" : "key",
          "in" : "query",
          "description" : "Your private API key.",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "domain",
          "in" : "query",
          "description" : "Integer value for the Amazon locale you want to access. (1-com, 2-co.uk, 3-de, 4-fr, 5-co.jp, 6-ca, 7-cn, 8-it, 9-es, 10-in, 11-com.mx, 12-com.br, 13-com.au)",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "enum" : [ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13 ]
          }
        }, {
          "name" : "category",
          "in" : "query",
          "description" : "The category node id of the category you want to request. For batch requests a comma separated list of ids (up to 10, the token cost stays the same). Alternatively you can specify the value 0, which will retrieve a list of all root categories.",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32"
          }
        }, {
          "name" : "parents",
          "in" : "query",
          "description" : "Whether or not to include the category tree for each category. (1 = include, 0 = do not include.)",
          "required" : true,
          "style" : "form",
          "explode" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "enum" : [ 0, 1 ]
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "search results matching criteria",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/category"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "bad input parameter"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "product" : {
        "required" : [ "asin", "code", "domainId" ],
        "type" : "object"
      },
      "category" : {
        "required" : [ "catId", "children", "domainId", "highestRank", "name", "parent", "productCount" ],
        "type" : "object",
        "properties" : {
          "domainId" : {
            "type" : "integer",
            "description" : "Integer value for the Amazon locale you want to access. (1-com, 2-co.uk, 3-de, 4-fr, 5-co.jp, 6-ca, 7-cn, 8-it, 9-es, 10-in, 11-com.mx, 12-com.br, 13-com.au)",
            "format" : "int32",
            "example" : 1,
            "enum" : [ 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13 ]
          },
          "catId" : {
            "type" : "integer",
            "description" : "The category node id used by Amazon. Represents the identifier of the category. Also part of the Product object’s categories and rootCategory fields. Always a positive Long value. Note - The id 9223372036854775807 (max signed long value) denotes a blank category with the name “?”. We use this in cases where a product is listed in no or non-existent categories.",
            "format" : "int32",
            "example" : 281052
          },
          "name" : {
            "type" : "string",
            "description" : "The name of the category.",
            "example" : "Digital Cameras"
          },
          "children" : {
            "type" : "array",
            "description" : "List of all sub categories. null or [] (empty array) if the category has no sub categories.",
            "example" : [ 3109924011, 7242008011, 3017941, 2476680011, 330405011, 2476681011, 3109925011 ],
            "items" : {
              "type" : "integer",
              "format" : "int32",
              "nullable" : true
            }
          },
          "parent" : {
            "type" : "integer",
            "description" : "The parent category’s Id. Always a positive Long value. If it is 0 the category is a root category and has no parent category.",
            "format" : "int32",
            "example" : 502394
          },
          "highestRank" : {
            "type" : "integer",
            "description" : "The highest (root category) sales rank we have observed of a product that is listed in this category. Note- Estimate, as the value is from the Keepa product database and not retrieved from Amazon.",
            "format" : "int32",
            "example" : 6549
          },
          "productCount" : {
            "type" : "integer",
            "description" : "Number of products that are listed in this category. Note- Estimate, as the value is from the Keepa product database and not retrieved from Amazon.",
            "format" : "int32",
            "example" : 1522
          }
        },
        "example" : {
          "catId" : 281052,
          "parent" : 502394,
          "children" : [ 3109924011, 7242008011, 3017941, 2476680011, 330405011, 2476681011, 3109925011 ],
          "highestRank" : 6549,
          "name" : "Digital Cameras",
          "productCount" : 1522,
          "domainId" : 1
        }
      }
    }
  }
}